Rename : 탭 컴포넌트 이름 변경
- product-detail-tab으로 변경
